[Metadata Integration] Renamed Seo to Metadata
