+++ src/Sylius/Bundle/CoreBundle/Resources/config/app/sylius.yml
@@ -744,7 +744,9 @@ sylius_user:
             model: Sylius\Component\Core\Model\Customer
             controller: Sylius\Bundle\CoreBundle\Controller\CustomerController
 
-sylius_seo:
+sylius_metadata:
     classes:
         metadata:
+            controller: Sylius\Bundle\WebBundle\Controller\Backend\MetadataController
+            repository: Sylius\Bundle\CoreBundle\Doctrine\ORM\MetadataRepository
             form: Sylius\Bundle\CoreBundle\Form\Type\RootMetadataFormType
