+++ src/Sylius/Bundle/WebBundle/Resources/translations/messages.en.yml
@@ -205,6 +205,8 @@ sylius:
             index: Menus
         menu_node:
             index: Menu nodes
+        metadata:
+            index: Metadatas
         new: New
         option:
             index: Options
@@ -501,6 +503,14 @@ sylius:
         frontend_description: Sylius is modern ecommerce solution for PHP, based on the Symfony2 framework.
         frontend_keywords: sylius, ecommerce, symfony2, shop, cart, webshop
         frontend_title: Sylius, modern ecommerce for Symfony2
+    metadata:
+        customize_header: Customizing metadata
+        show_header: Show metadata
+        id: ID
+        customize: Customize metadata
+        manage: Manage metadata
+        content: Content
+        index_metadata: Metadatas
     na: N/A
     newest: Newest products
     no: no
