+++ src/Sylius/Bundle/WebBundle/Resources/views/Backend/Metadata/_form.html.twig
@@ -5,8 +5,14 @@
 <fieldset id="metadata-form">
     <div class="row">
         <div class="col-md-6">
-            {{ form_row(form.key) }}
-            {{ form_row(form.metadata) }}
+            {{ form_row(form.id) }}
+            <hr/>
+            {{ form_row(form.metadata.title) }}
+            {{ form_row(form.metadata.description) }}
+            {{ form_row(form.metadata.keywords) }}
+        </div>
+        <div class="col-md-6">
+            {{ form_row(form.metadata.twitter) }}
         </div>
     </div>
 </fieldset>
