--- src/Sylius/Bundle/WebBundle/Resources/views/Backend/Metadata/create.html.twig
@@ -1,29 +1,0 @@
-{% extends 'SyliusWebBundle:Backend:layout.html.twig' %}
-
-{% from 'SyliusResourceBundle:Macros:actions.html.twig' import create %}
-
-{% block javascripts %}
-    {{ parent() }}
-    <script type="text/javascript" src="{{ asset('bundles/syliusseo/js/dynamic-form.js') }}"></script>
-{% endblock %}
-
-{% block topbar %}
-<ol class="breadcrumb">
-    <li>{{ 'sylius.breadcrumb.configuration'|trans }}</li>
-    <li><a href="{{ path('sylius_backend_metadata_index') }}">{{ 'sylius.breadcrumb.metadata.index'|trans }}</a></li>
-    <li>{{ 'sylius.breadcrumb.new'|trans }}</li>
-</ol>
-{% endblock %}
-
-{% block content %}
-<div class="page-header">
-    <h1><i class="glyphicon glyphicon-plus-sign"></i> {{ 'sylius.metadata.create_header'|trans|raw }}</h1>
-</div>
-
-{{ form_errors(form) }}
-
-<form action="{{ path('sylius_backend_metadata_create') }}" method="post" class="form-horizontal" novalidate>
-    {% include 'SyliusWebBundle:Backend/Metadata:_form.html.twig' %}
-    {{ create() }}
-</form>
-{% endblock %}
