+++ src/Sylius/Bundle/WebBundle/Resources/views/Backend/Metadata/customize.html.twig
@@ -4,27 +4,26 @@
 
 {% block javascripts %}
     {{ parent() }}
-    <script type="text/javascript" src="{{ asset('bundles/syliusseo/js/dynamic-form.js') }}"></script>
+    <script type="text/javascript" src="{{ asset('bundles/syliusmetadata/js/dynamic-form.js') }}"></script>
 {% endblock %}
 
 {% block topbar %}
 <ol class="breadcrumb">
     <li>{{ 'sylius.breadcrumb.configuration'|trans }}</li>
     <li><a href="{{ path('sylius_backend_metadata_index') }}">{{ 'sylius.breadcrumb.metadata.index'|trans }}</a></li>
-    <li><a href="{{ path('sylius_backend_metadata_show', {'id': metadata.id}) }}">{{ metadata.key }}</a></li>
+    <li><a href="{{ path('sylius_backend_metadata_show', {'id': metadata.id}) }}">{{ metadata.id }}</a></li>
     <li>{{ 'sylius.breadcrumb.edit'|trans }}</li>
 </ol>
 {% endblock %}
 
 {% block content %}
 <div class="page-header">
-    <h1><i class="glyphicon glyphicon-pencil"></i> {{ 'sylius.metadata.update_header'|trans|raw }}</h1>
+    <h1><i class="glyphicon glyphicon-pencil"></i> {{ 'sylius.metadata.customize_header'|trans|raw }}</h1>
 </div>
 
 {{ form_errors(form) }}
 
-<form action="{{ path('sylius_backend_metadata_update', {'id': metadata.id}) }}" method="post" class="form-horizontal" novalidate>
-    <input type="hidden" name="_method" value="PUT">
+<form action="{{ path('sylius_backend_metadata_customize', {'id': metadata.id }) }}" method="post" class="form-horizontal" novalidate>
     {% include 'SyliusWebBundle:Backend/Metadata:_form.html.twig' %}
     {{ update() }}
 </form>
