+++ src/Sylius/Component/Core/Model/ProductVariantInterface.php
@@ -15,7 +15,7 @@ use Doctrine\Common\Collections\Collection;
 use Sylius\Component\Inventory\Model\StockableInterface;
 use Sylius\Component\Pricing\Model\PriceableInterface;
 use Sylius\Component\Product\Model\VariantInterface as BaseVariantInterface;
-use Sylius\Component\Seo\Model\MetadataSubjectInterface;
+use Sylius\Component\Metadata\Model\MetadataSubjectInterface;
 use Sylius\Component\Shipping\Model\ShippableInterface;
 
 /**
