+++ src/Sylius/Component/Core/spec/Metadata/HierarchyProvider/ProductHierarchyProviderSpec.php
@@ -14,7 +14,7 @@ namespace spec\Sylius\Component\Core\Metadata\HierarchyProvider;
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Core\Model\ArchetypeInterface;
 use Sylius\Component\Core\Model\ProductInterface;
-use Sylius\Component\Seo\Model\MetadataSubjectInterface;
+use Sylius\Component\Metadata\Model\MetadataSubjectInterface;
 
 /**
  * @mixin \Sylius\Component\Core\Metadata\HierarchyProvider\ProductHierarchyProvider
@@ -28,9 +28,9 @@ class ProductHierarchyProviderSpec extends ObjectBehavior
         $this->shouldHaveType('Sylius\Component\Core\Metadata\HierarchyProvider\ProductHierarchyProvider');
     }
 
-    function it_implements_Sylius_Seo_Metadata_Hierarchy_Provider_interface()
+    function it_implements_Sylius_Metadata_Hierarchy_Provider_interface()
     {
-        $this->shouldImplement('Sylius\Component\Seo\HierarchyProvider\MetadataHierarchyProviderInterface');
+        $this->shouldImplement('Sylius\Component\Metadata\HierarchyProvider\MetadataHierarchyProviderInterface');
     }
 
     function it_supports_Sylius_Core_Product(
